# gh-100726: optimize construction of range object for medium sized integers