--- a/Misc/NEWS.d/next/Core and Builtins/2023-01-03-20-59-20.gh-issue-100726.W9huFl.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2023-01-03-20-59-20.gh-issue-100726.W9huFl.rst
@@ -1,1 +1,1 @@
-Optimize construction of `range` object for medium size integers.
+Optimize construction of ``range`` object for medium size integers.