--- a/Objects/rangeobject.c
+++ b/Objects/rangeobject.c
@@ -177,6 +177,7 @@ get_len_of_range(long lo, long hi, long step);
 /// Return the length as a long, or -1 on error
 long compute_range_length_long(PyLongObject *start, PyLongObject *stop, PyLongObject *step) {
     int overflow = 0;
+
     long long_start = _PyLong_AsLongAndOverflow(start, &overflow);
     if (overflow) {
             return -1;