The python `range` object contains code to compute the length from the start, stop and step values. This calculation is performed with PyLong objects which is expensive. This PR adds a fast path for the common case where start, stop and step all fit into a long value.

Benchmark results:
```
range(0, 5): Mean +- std dev: [main_range] 110 ns +- 1 ns -> [pr_range] 56.8 ns +- 1.3 ns: 1.93x faster
list(range(0, 5)): Mean +- std dev: [main_range] 243 ns +- 1 ns -> [pr_range] 181 ns +- 3 ns: 1.34x faster
range(-2, 3): Mean +- std dev: [main_range] 110 ns +- 2 ns -> [pr_range] 56.6 ns +- 1.0 ns: 1.94x faster
list(range(-2, 3)): Mean +- std dev: [main_range] 244 ns +- 3 ns -> [pr_range] 182 ns +- 4 ns: 1.34x faster
range(2, 60): Mean +- std dev: [main_range] 112 ns +- 1 ns -> [pr_range] 57.1 ns +- 1.4 ns: 1.96x faster
list(range(2, 60)): Mean +- std dev: [main_range] 535 ns +- 28 ns -> [pr_range] 472 ns +- 27 ns: 1.13x faster
range(2, 60, 2): Mean +- std dev: [main_range] 115 ns +- 1 ns -> [pr_range] 59.4 ns +- 1.4 ns: 1.94x faster
list(range(2, 60, 2)): Mean +- std dev: [main_range] 382 ns +- 11 ns -> [pr_range] 317 ns +- 5 ns: 1.21x faster
range(73786976294838206464, 73786976294838206469, 2): Mean +- std dev: [main_range] 136 ns +- 1 ns -> [pr_range] 137 ns +- 3 ns: 1.01x slower
list(range(73786976294838206464, 73786976294838206469, 2)): Mean +- std dev: [main_range] 444 ns +- 3 ns -> [pr_range] 440 ns +- 6 ns: 1.01x faster
for loop with range): Mean +- std dev: [main_range] 878 ns +- 12 ns -> [pr_range] 815 ns +- 21 ns: 1.08x faster

Geometric mean: 1.39x faster
```

<details>
<summary>Benchmark script</summary>

```
import pyperf
runner = pyperf.Runner()

cases = [(0,5), (-2, 3), (2, 60)]
for s,t in cases:
    time = runner.timeit(name=f'range({s}, {t})', stmt=f"range({s}, {t})")
    time = runner.timeit(name=f'list(range({s}, {t}))', stmt=f"list(range({s}, {t}))")

time = runner.timeit(name=f'range({s}, {t}, 2)', stmt=f"range({s}, {t}, 2)")
time = runner.timeit(name=f'list(range({s}, {t}, 2))', stmt=f"list(range({s}, {t}, 2))")

big = 2**66
s,t=big, big+5
time = runner.timeit(name=f'range({s}, {t}, 2)', stmt=f"range({s}, {t}, 2)")
time = runner.timeit(name=f'list(range({s}, {t}, 2))', stmt=f"list(range({s}, {t}, 2))")
    
stmt="""
v=0
for ii in range(20):
    v += ii*ii
"""
time = runner.timeit(name='for loop with range)', stmt=stmt)
```
</details>

**Notes**

* The macro and inline method `IS_MEDIUM_VALUE` and `medium_value` have been moved from `longobject.c` to the internal header files. The names have been kept the same to avoid too many changes. But since they are now visible outside `longobject.c` we can rename them to `_PyLong_IsMediumValue` and `_PyLong_MediumValue`.


<!-- gh-issue-number: gh-100726 -->
* Issue: gh-100726
<!-- /gh-issue-number -->