@mdickinson 2023-01-04
Could you try this using `PyLong_AsLong` instead of `MEDIUM_VALUE`? This would cover more cases, be more consistent with the existing long-handling in `rangeobject.c`, and avoid the internal details of the long object representation leaking out too much. I'm not too comfortable with making the range implementation dependent on the longobject internals.