@eendebakpt 2023-01-04
> Could you try this using `PyLong_AsLong` instead of `MEDIUM_VALUE`? This would cover more cases, be more consistent with the existing long-handling in `rangeobject.c`, and avoid the internal details of the long object representation leaking out too much. I'm not too comfortable with making the range implementation dependent on the longobject internals.

@mdickinson The `PyLong_AsLong` indeed has many advantages, but it internally sets error values and needs `PyErr_Occurred()` to disambiguate  which impacts performance. 

The `PyLong_AsLongAndOverflow` is also public and perform a bit less work, so I tried that method. The [branch with `PyLong_AsLongAndOverflow`](https://github.com/python/cpython/compare/main...eendebakpt:cpython:range_fast_path_v2) is also faster than main for the common case (but not as fast as this PR). For the case of large integers `PyLong_AsLongAndOverflow` is slower than main (but not a lot).

**Benchmark results**

Main:
```
range(0, 5): Mean +- std dev: 110 ns +- 1 ns
list(range(0, 5)): Mean +- std dev: 243 ns +- 1 ns
range(-2, 3): Mean +- std dev: 110 ns +- 2 ns
list(range(-2, 3)): Mean +- std dev: 244 ns +- 3 ns
range(2, 60): Mean +- std dev: 112 ns +- 1 ns
list(range(2, 60)): Mean +- std dev: 535 ns +- 28 ns
range(2, 60, 2): Mean +- std dev: 115 ns +- 1 ns
list(range(2, 60, 2)): Mean +- std dev: 382 ns +- 11 ns
range(73786976294838206464, 73786976294838206469, 2): Mean +- std dev: 136 ns +- 1 ns
list(range(73786976294838206464, 73786976294838206469, 2)): Mean +- std dev: 444 ns +- 3 ns
for loop with range(20): Mean +- std dev: 878 ns +- 12 ns
```

This PR:
```
range(0, 5): Mean +- std dev: 56.8 ns +- 1.3 ns
list(range(0, 5)): Mean +- std dev: 181 ns +- 3 ns
range(-2, 3): Mean +- std dev: 56.6 ns +- 1.0 ns
list(range(-2, 3)): Mean +- std dev: 182 ns +- 4 ns
range(2, 60): Mean +- std dev: 57.1 ns +- 1.4 ns
list(range(2, 60)): Mean +- std dev: 472 ns +- 27 ns
range(2, 60, 2): Mean +- std dev: 59.4 ns +- 1.4 ns
list(range(2, 60, 2)): Mean +- std dev: 317 ns +- 5 ns
range(73786976294838206464, 73786976294838206469, 2): Mean +- std dev: 137 ns +- 3 ns
list(range(73786976294838206464, 73786976294838206469, 2)): Mean +- std dev: 440 ns +- 6 ns
for loop with range(20): Mean +- std dev: 815 ns +- 21 ns
```

Branch using `PyLong_AsLongAndOverflow`: https://github.com/python/cpython/compare/main...eendebakpt:cpython:range_fast_path_v2
```
range(0, 5): Mean +- std dev: 63.3 ns +- 0.8 ns
list(range(0, 5)): Mean +- std dev: 193 ns +- 3 ns
range(-2, 3): Mean +- std dev: 63.9 ns +- 1.3 ns
list(range(-2, 3)): Mean +- std dev: 194 ns +- 3 ns
range(2, 60): Mean +- std dev: 64.0 ns +- 1.7 ns
list(range(2, 60)): Mean +- std dev: 478 ns +- 10 ns
range(2, 60, 2): Mean +- std dev: 66.8 ns +- 3.5 ns
list(range(2, 60, 2)): Mean +- std dev: 333 ns +- 13 ns
range(73786976294838206464, 73786976294838206469, 2): Mean +- std dev: 149 ns +- 6 ns
list(range(73786976294838206464, 73786976294838206469, 2)): Mean +- std dev: 455 ns +- 8 ns
for loop with range(20): Mean +- std dev: 831 ns +- 17 ns
```

The benchmark of `list(range(73786976294838206464, 73786976294838206469, 2))` slows down from main to the `PyLong_AsLongOverflow` from 444 to 455 ns. (`range(73786976294838206464, 73786976294838206469, 2)` slows down as well, but a standalone `range` is hardly used).

I am fine with accepting a minor slowdown for the uncommon case, although perhaps there are more cases to consider than I included in the benchmark.

Which cases would `PyLong_AsLong` or `PyLong_AsLongOverflow` cover more than the `IS_MEDIUM_VALUE`? According to the documentation the input arguments to `compute_range_length` should satisfy `PyLong_Check(arg)=1`. So values of like `numpy.int64(3)` or `Decimal(4,1)` are converted at an earlier stage (e.g. in `range_from_array`).