@mdickinson 2023-01-04
> Which cases would `PyLong_AsLong` or `PyLong_AsLongOverflow` cover more than the `IS_MEDIUM_VALUE`?

Using a `PyLong_AsLong` (or a variant) would allow anything that fits in a C `long` to use the fast path, rather than just those values that fit in a single digit. So on a typical machine, that's all values between `-2**63` and `2**63-1`, rather than just values in the range `[-2**30, 2**30]` (the "medium" values).

IOW, thinking of the range implementation as a client user of the PyLong implementation, what that client needs is a way to say, efficiently, "tell me whether this object that I already know is a PyLongObject fits in a C long, and if so, give me that value". `IS_MEDIUM_VALUE` is doing something quite different, that's dependent on the internal representation of a Python long - the range implementation shouldn't know or care about "medium" values.

It's that information about digits, twodigits, medium values, number of bits per digit, etc., that I think we should keep encapsulated within the long implementation - it shouldn't be relevant to users of that implementation. If it starts leaking into other parts of the codebase, that's going to turn into a maintenance headache whenever the long implementation needs tweaking.

If none of the existing public or private long API functions does what we need efficiently (e.g., because it's doing index checks and the like), I'd be open to adding a new function that does answer the "tell me whether this PyLong fits in a C long" question efficiently.