@eendebakpt 2023-01-04
@mdickinson I refactored the `PyLong_AsLongAndOverflow` into a part with error checking and a new (private) `_PyLong_AsLongAndOverflow` that performs the actual conversion to `long` and overflow check.

Benchmark of the updated PR against main:
```
range(0, 5): Mean +- std dev: [main] 110 ns +- 1 ns -> [pr_v4] 61.2 ns +- 1.5 ns: 1.79x faster
list(range(0, 5)): Mean +- std dev: [main] 244 ns +- 2 ns -> [pr_v4] 186 ns +- 1 ns: 1.31x faster
range(-2, 3): Mean +- std dev: [main] 111 ns +- 4 ns -> [pr_v4] 61.6 ns +- 1.5 ns: 1.80x faster
list(range(-2, 3)): Mean +- std dev: [main] 244 ns +- 1 ns -> [pr_v4] 187 ns +- 1 ns: 1.30x faster
range(2, 60): Mean +- std dev: [main] 112 ns +- 1 ns -> [pr_v4] 62.3 ns +- 3.7 ns: 1.79x faster
list(range(2, 60)): Mean +- std dev: [main] 527 ns +- 7 ns -> [pr_v4] 466 ns +- 9 ns: 1.13x faster
range(2, 60, 2): Mean +- std dev: [main] 116 ns +- 1 ns -> [pr_v4] 64.2 ns +- 1.0 ns: 1.80x faster
list(range(2, 60, 2)): Mean +- std dev: [main] 379 ns +- 3 ns -> [pr_v4] 321 ns +- 8 ns: 1.18x faster
range(4398046511104, 4398046511109, 2): Mean +- std dev: [main] 135 ns +- 1 ns -> [pr_v4] 68.1 ns +- 1.9 ns: 1.98x faster
list(range(4398046511104, 4398046511109, 2)): Mean +- std dev: [main] 311 ns +- 4 ns -> [pr_v4] 235 ns +- 4 ns: 1.32x faster
range(73786976294838206464, 73786976294838206469, 2): Mean +- std dev: [main] 137 ns +- 6 ns -> [pr_v4] 148 ns +- 1 ns: 1.08x slower
list(range(73786976294838206464, 73786976294838206469, 2)): Mean +- std dev: [main] 444 ns +- 2 ns -> [pr_v4] 447 ns +- 12 ns: 1.01x slower
for loop with range(10): Mean +- std dev: [main] 879 ns +- 17 ns -> [pr_v4] 826 ns +- 16 ns: 1.06x faster
for loop with range(20): Mean +- std dev: [main] 881 ns +- 22 ns -> [pr_v4] 824 ns +- 19 ns: 1.07x faster

Geometric mean: 1.35x faster
```
Most test cases are faster, now also the cases with long values that are not medium. The new check `_PyLong_AsLongAndOverflow` is a bit slower than `IS_MEDIUM_INT`, because it has to do a bit more work.